Simplified code to produce less files with same info
I prefer this version of the script over the 'Read from Group' version. As this version, pulls the device names from a list and queries Azure for the info, and in turn is able to spot duplicates, which the Read from Group version is not able to do. Therefore this is the copy I will be updating and work with moving forward.